* configure.ac: * src/bin/Makefile.am: * src/lib/Makefile.am: * src/lib/dlfcn/Makefile.am: * src/lib/mman/Makefile.am: use -Wl,--enable-auto-import with cegcc fix warning with newest autoconf * src/lib/dlfcn/dlfcn.c: (string_to_wchar): include wchar.h and fix warnings * src/lib/mman/mman.c: use relative path to include sys/mman.h
SVN revision: 34348

+++ legacy/evil/src/lib/dlfcn/dlfcn.c
@@ -5,6 +5,7 @@
 #undef WIN32_LEAN_AND_MEAN
 
 #if defined(__CEGCC__) || defined(__MINGW32CE__)
+# include <wchar.h>
 # include <limits.h>
 #endif /* __MINGW32CE__ */
 
@@ -31,7 +32,8 @@ string_to_wchar(const char *text)
    int      wsize;
 
    wsize = MultiByteToWideChar(CP_ACP, 0, text, strlen(text) + 1, NULL, 0);
-   if (wsize == 0 || wsize > ULONG_MAX/sizeof(wchar_t))
+   if ((wsize == 0) ||
+       (wsize > (int)(ULONG_MAX/sizeof(wchar_t))))
      return NULL;
 
    wtext = malloc(wsize * sizeof(wchar_t));
